sms verification: move server-related code to 3rdparty
