+++ js/src/modals/SMSVerification/store.js
@@ -20,7 +20,8 @@ import { sha3 } from '../../api/util/sha3';
 
 import Contracts from '../../contracts';
 
-import { checkIfVerified, checkIfRequested, postToServer } from '../../contracts/sms-verification';
+import { checkIfVerified, checkIfRequested } from '../../contracts/sms-verification';
+import { postToServer } from '../../3rdparty/sms-verification';
 import checkIfTxFailed from '../../util/check-if-tx-failed';
 import waitForConfirmations from '../../util/wait-for-block-confirmations';
